requirements.txt
# Make sure to install the torch dependencies with conda:
# conda install pytorch=1.11.0 torchdata torchvision -c pytorch -y

fancy_einsum
einops
requests
transformers
pytest
click
joblib
tqdm
ipykernel
notebook
ipywidgets
jupyter
flask
matplotlib
numpy-stl
mediapy
pandas
scikit-learn
black
wandb
sentence-transformers
ftfy
tensorboard
pygame
plotly
sympy
accelerate
pre-commit
diffusers
torchtyping
typeguard
streamlit
torchinfo
frozendict
streamlit-chat
openai